requirements.txt
langgraph
langchain
langchain_core
langchain_google_genai
faiss
keras
tensorflow
sentence_transformers
tqdm
numpy
fitz
pinecone
fastapi
uvicorn
pydantic
requests
python-docx
python-dotenv
aiofiles
